requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.12.0=py39h06a4308_0
aiohttp=3.7.4=py39h27cfd23_1
argparse=1.4.0=pypi_0
astunparse=1.6.3=py_0
async-timeout=3.0.1=py39h06a4308_0
attrs=20.3.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py39h06a4308_0
brotlipy=0.7.0=py39h27cfd23_1003
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.1.19=h06a4308_1
cachetools=4.2.1=pyhd3eb1b0_0
certifi=2020.12.5=py39h06a4308_0
cffi=1.14.5=py39h261ae71_0
chardet=3.0.4=py39h06a4308_1003
click=7.1.2=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
coverage=5.5=py39h27cfd23_2
crayons=0.3.1=pypi_0
cryptography=2.9.2=py39h3c74f83_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cycler=0.10.0=py39h06a4308_0
cython=0.29.22=py39h2531618_0
dbus=1.13.18=hb2f20db_0
essential-generators=1.0=pypi_0
expat=2.3.0=h2531618_2
fontconfig=2.13.1=h6c09931_0
freetype=2.10.4=h5ab3b9f_0
gast=0.4.0=py_0
glib=2.68.0=h36276a3_0
google-auth=1.27.1=pyhd3eb1b0_0
google-auth-oauthlib=0.4.2=pyhd3eb1b0_2
google-pasta=0.2.0=py_0
grpcio=1.36.1=py39h2157cd5_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
h5py=2.10.0=py39hec9cf62_0
hdf5=1.10.6=hb1b8bf9_0
icu=58.2=he6710b0_3
idna=2.10=pyhd3eb1b0_0
importlib-metadata=3.7.3=py39h06a4308_1
intel-openmp=2020.2=254
joblib=1.0.1=pyhd3eb1b0_0
jpeg=9b=h024ee3a_2
keras-base=2.4.3=py_0
keras-gpu=2.4.3=0
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.1=py39h2531618_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.14.0=h8c45485_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_1
libuuid=1.0.3=h1bed415_2
libxcb=1.14=h7b6447c_0
libxml2=2.9.10=hb55368b_3
lz4-c=1.9.3=h2531618_0
markdown=3.3.4=py39h06a4308_0
matplotlib=3.3.4=py39h06a4308_0
matplotlib-base=3.3.4=py39h62a2d02_0
mkl=2020.2=256
mkl-service=2.3.0=py39he8ac12f_0
mkl_fft=1.3.0=py39h54f3939_0
mkl_random=1.0.2=py39h63df603_0
multidict=5.1.0=py39h27cfd23_2
ncurses=6.2=he6710b0_1
numpy=1.19.2=py39h89c1606_0
numpy-base=1.19.2=py39h2ae0177_0
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openssl=1.1.1k=h27cfd23_0
opt_einsum=3.1.0=py_0
pcre=8.44=he6710b0_0
piexif=1.1.3=pypi_0
pillow=6.2.2=pypi_0
pip=21.0.1=py39h06a4308_0
protobuf=3.14.0=py39h2531618_1
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=2.0.1=py39h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pyqt=5.9.2=py39h2531618_6
pysocks=1.7.1=py39h06a4308_0
python=3.9.2=hdb3f193_0
python-dateutil=2.8.1=pyhd3eb1b0_0
python-flatbuffers=1.12=pyhd3eb1b0_0
pyyaml=5.4.1=py39h27cfd23_1
qt=5.9.7=h5867ecd_1
readline=8.1=h27cfd23_0
requests=2.25.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=0.24.1=py39ha9443f7_0
scipy=1.6.1=py39h91f5cce_0
setuptools=52.0.0=py39h06a4308_0
sip=4.19.13=py39h2531618_0
six=1.15.0=py39h06a4308_0
sqlite=3.35.1=hdfb4753_0
stegano=0.9.8=pypi_0
tensorboard=2.4.0=pyhc547734_0
tensorboard-plugin-wit=1.6.0=py_0
tensorflow=2.4.1=gpu_py39h8236f22_0
tensorflow-base=2.4.1=gpu_py39h29c2da4_0
tensorflow-estimator=2.4.1=pyheb71bc4_0
tensorflow-gpu=2.4.1=h30adc30_0
termcolor=1.1.0=py39h06a4308_1
threadpoolctl=2.1.0=pyh5ca1d4c_0
tk=8.6.10=hbc83047_0
tornado=6.1=py39h27cfd23_0
tqdm=4.59.0=pypi_0
typing-extensions=3.7.4.3=hd3eb1b0_0
typing_extensions=3.7.4.3=pyh06a4308_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.4=pyhd3eb1b0_0
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=py39he8ac12f_1
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py39h27cfd23_0
zipp=3.4.1=pyhd3eb1b0_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
